backend/rag-vectorize/requirements.txt
psycopg2-binary==2.9.9
openai==1.54.0
httpx==0.27.0